Updated structure and styling
