+++ src/myapp/client/src/Components/Header.js
@@ -1,5 +1,5 @@
 import React from "react"
-import './App.css';
+import '../Styles/App.css';
 
 function Header(){
     return(
